+++ config/.env.base.sh
@@ -25,8 +25,6 @@ export METRICS_SERVICE='disabled'
 export MATOMO_TRACKER_URL='disabled'
 export MATOMO_TRACKER_SITEID='0'
 export SUPPORT_WEB_ADDRESS='https://example.com/'
-# Crowdin translation service (enabled / disabled):
-export CROWDIN='disabled'
 # Stonly service https://stonly.com/ (enabled / disabled):
 export STONLY_WID='disabled'
 # axeptio.eu cookie banner