+++ templates/crowdin.js
@@ -1,3 +1,0 @@
-/* Crowdin configuration */
-const _jipt = []; // eslint-disable-line no-underscore-dangle
-_jipt.push(['project', 'BRANDING']);