+++ CHANGELOG.md
@@ -14,6 +14,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1042 Bump some dependencies
 - YLP-1049 Bump node to v14
 - YLP-1051 Allow to verify with eslint js files outside main packages
+- YLP-1053 Remove lolcat language and Crowdin build
 
 ## 2.0.4 - 2021-09-27
 ### Added